+++ shaungc-java-dev/src/main/java/com/shaungc/utilities/RedisPubSubChannelName.java
@@ -1,0 +1,15 @@
+package com.shaungc.utilities;
+
+public enum RedisPubSubChannelName {
+    SCRAPER_JOB_CHANNEL("scraperJobChannel");
+
+    private final String string;
+
+    private RedisPubSubChannelName(String string) {
+        this.string = string;
+    }
+
+    public String getString() {
+        return this.string;
+    }
+}
