+++ shaungc-java-dev/src/main/java/com/shaungc/utilities/ScraperJobMessageTo.java
@@ -1,0 +1,16 @@
+package com.shaungc.utilities;
+
+public enum ScraperJobMessageTo {
+    SLACK_MD_SVC("slackMiddlewareService"),
+    SCRAPER("scraper");
+
+    private final String string;
+
+    private ScraperJobMessageTo(String string) {
+        this.string = string;
+    }
+
+    public String getString() {
+        return this.string;
+    }
+}
