+++ shaungc-java-dev/src/main/java/com/shaungc/utilities/ScraperJobMessageType.java
@@ -1,0 +1,18 @@
+package com.shaungc.utilities;
+
+public enum ScraperJobMessageType {
+    PREFLIGHT("preflight"),
+    PROGRESS("progress"),
+    FINISH("finish"),
+    ERROR("error");
+
+    private final String string;
+
+    private ScraperJobMessageType(String string) {
+        this.string = string;
+    }
+
+    public String getString() {
+        return this.string;
+    }
+}
